+++ controller/registerController.js
@@ -1,6 +1,6 @@
 const User = require("../model/User");
 const bcrypt = require("bcrypt");
-
+const {createJWT} = require("../controller/createJWT")
 const handleNewUser = async (req, res) => {
   //이메일, 유저네임, 패스워드
   const { mail, name, pwd } = req.body;
@@ -22,7 +22,8 @@ const handleNewUser = async (req, res) => {
     });
 
     console.log(newUser);
-    res.status(200).json({ success: true });
+    createJWT(res,newUser);
+    //res.status(200).json({ success: true });
   } catch (err) {
     res.status(500).json({ message: err.message });
   }
